requirements.txt
praw==7.7.1
httpx==0.27.2
pydantic==2.9.2
pyyaml==6.0.2
rich==13.9.2
python-dotenv==1.0.1
aiolimiter==1.1.0
sqlite-utils==3.37
